Fix gc/readweakvaltable races
The following pattern leads to wrong behavior.
```
weakvaltable[a] = function(val) -- 1
<GC TRIGGERED>
return weakvaltable[a] -- 2
```
In case GC is triggered between lines 1 and 2,
the retval would be equal to `nil`, because there are
no reverences to the value between those lines.

This commit fixes such patterns in the code with help
of intermediate variable.